# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      timestamp:
        type: string
    type: object
  model.AliasRequest:
    properties:
      alias_name:
        maxLength: 100
        minLength: 1
        type: string
      item_id:
        type: integer
      name_type:
        allOf:
        - $ref: '#/definitions/model.NameType'
        enum:
        - COMMON
        - ALIAS
      region_id:
        type: integer
    required:
    - alias_name
    - item_id
    - name_type
    - region_id
    type: object
  model.AliasResponse:
    properties:
      alias_name:
        type: string
      created_at:
        type: string
      id:
        type: integer
      item_id:
        type: integer
      name_type:
        $ref: '#/definitions/model.NameType'
      region_id:
        type: integer
      reviewed_at:
        type: string
      status:
        $ref: '#/definitions/model.AliasStatus'
      submitted_by:
        type: integer
      votes_count:
        type: integer
    type: object
  model.AliasStatus:
    enum:
    - PENDING
    - APPROVED
    - REJECTED
    type: string
    x-enum-varnames:
    - AliasStatusPending
    - AliasStatusApproved
    - AliasStatusRejected
  model.AnonymousAliasRequest:
    properties:
      alias_name:
        maxLength: 100
        minLength: 1
        type: string
      item_id:
        type: integer
      name_type:
        allOf:
        - $ref: '#/definitions/model.NameType'
        enum:
        - COMMON
        - ALIAS
      region_id:
        type: integer
      submitter:
        description: 提交者昵称
        maxLength: 100
        minLength: 2
        type: string
    required:
    - alias_name
    - item_id
    - name_type
    - region_id
    - submitter
    type: object
  model.CategoryResponse:
    properties:
      children:
        items:
          $ref: '#/definitions/model.CategoryResponse'
        type: array
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      sort_order:
        type: integer
    type: object
  model.NameType:
    enum:
    - COMMON
    - ALIAS
    type: string
    x-enum-varnames:
    - NameTypeCommon
    - NameTypeAlias
...
  model.UserLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  model.UserRegisterRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  model.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_verified:
        type: boolean
      role:
        type: string
      username:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: support@cnalias.com
    name: API Support
  description: 中国地区叫法对比平台 API - 记录中国不同地区对同一物品的不同叫法（方言/别名）
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://cnalias.com/terms/
  title: cnalias API
  version: 1.0.0
paths:
...
  /aliases:
    post:
      consumes:
      - application/json
      description: Submit a new regional alias for an item (requires authentication)
      parameters:
      - description: Alias submission
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.AliasRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AliasResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Submit a new alias
      tags:
      - aliases
  /aliases/anonymous:
    post:
      consumes:
      - application/json
      description: Submit a new regional alias without authentication (requires review)
      parameters:
      - description: Anonymous alias submission
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.AnonymousAliasRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AliasResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Submit alias anonymously
      tags:
      - aliases
  /aliases/my-submissions:
    get:
      consumes:
      - application/json
      description: Get list of aliases submitted by the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.AliasResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Get user's submissions
      tags:
      - aliases
  /aliases/search:
    get:
      consumes:
      - application/json
      description: Search for items and aliases by keyword
      parameters:
      - description: Search keyword
        in: query
        name: q
        required: true
        type: string
      - description: Filter by region ID
        format: int64
        in: query
        name: region_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Search aliases
      tags:
      - aliases
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password to get access tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
      summary: User login
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Register a new user
      tags:
      - auth
  /categories:
    get:
      consumes:
      - application/json
      description: Get list of categories with optional parent filter
      parameters:
      - description: Filter by parent category ID
        format: int64
        in: query
        name: parent_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.CategoryResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: List categories
      tags:
      - categories
  /categories/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific category
      parameters:
      - description: Category ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.CategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Get category by ID
      tags:
      - categories
  /categories/tree:
    get:
      consumes:
      - application/json
      description: Get hierarchical category tree structure
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.CategoryResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Get category tree
      tags:
      - categories
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"